requirements.txt
pydub
moviepy
openai
git+https://github.com/openai/whisper.git
pyannote.audio
yt-dlp
spacy
deplacy
google-cloud-translate==2.0.1
six==1.15.0
python-dotenv
